+++ package.json
@@ -30,7 +30,6 @@
     "@babel/node": "^7.5.5",
     "@babel/plugin-proposal-class-properties": "^7.5.5",
     "@babel/preset-env": "^7.5.5",
-    "@hapi/joi": "^15.1.0",
     "babel-eslint": "^10.0.2",
     "body-parser": "^1.19.0",
     "chai": "^4.2.0",
@@ -41,6 +40,7 @@
     "env": "0.0.2",
     "express": "^4.17.1",
     "faker": "^4.1.0",
+    "joi": "^14.3.1",
     "js-yaml": "^3.13.1",
     "jsonwebtoken": "^8.5.1",
     "lodash": "^4.17.15",
