+++ server/routes/session_route.js
@@ -4,6 +4,8 @@ import ReviewController from '../controllers/review_controller';
 import auth from '../middleware/auth';
 import mentor from '../middleware/mentor';
 import admin from '../middleware/admin';
+import sessionValidate from '../middleware/sessionsValidator';
+import reviewValidate from '../middleware/reviewValidator';
 
 const router = express.Router();
 
@@ -11,7 +13,7 @@ const router = express.Router();
 const session_controller = new SessionController();
 const review_controller = new ReviewController();
 
-router.post('/sessions', auth, session_controller.create);
+router.post('/sessions', auth, sessionValidate, session_controller.create);
 
 router.patch('/sessions/:id/accept', mentor, session_controller.acceptSession);
 
@@ -19,7 +21,7 @@ router.patch('/sessions/:id/reject', mentor, session_controller.rejectSession);
 
 router.get('/sessions', auth, session_controller.yourOwnSessions);
 
-router.post('/sessions/:id/review', auth, review_controller.creatReview);
+router.post('/sessions/:id/review', auth, reviewValidate, review_controller.creatReview);
 
 router.delete('/sessions/:id/review', admin, review_controller.deleteReview);
