+++ server/routes/user_route.js
@@ -1,5 +1,6 @@
 import express from 'express';
 import UserController from '../controllers/user_controller';
+import userValidate from '../middleware/userValidator';
 
 
 const router = express.Router();
@@ -8,7 +9,7 @@ const router = express.Router();
 // creation of object
 const user_controller = new UserController();
 
-router.post('/signup', user_controller.signUp);
+router.post('/signup', userValidate, user_controller.signUp);
 
 // login user endpoint
 router.post('/signin', user_controller.signIn);
